0.1.3版本优化

+++ src/main/java/com/xiuxian/game/dto/request/ItemRemoveRequest.java
@@ -8,7 +8,7 @@ import lombok.Data;
 public class ItemRemoveRequest {
     @NotNull(message = "玩家物品ID不能为空")
     @Min(value = 1, message = "玩家物品ID必须大于0")
-    private Long playerItemId;
+    private Integer playerItemId;
 
     @NotNull(message = "数量不能为空")
     @Min(value = 1, message = "数量必须大于0")
