+++ src/main/java/com/xiuxian/game/dto/request/ItemUseRequest.java
@@ -8,5 +8,5 @@ import lombok.Data;
 public class ItemUseRequest {
     @NotNull(message = "玩家物品ID不能为空")
     @Min(value = 1, message = "玩家物品ID必须大于0")
-    private Long playerItemId;
+    private Integer playerItemId;
 }
