+++ src/main/java/com/xiuxian/game/dto/request/SkillUpgradeRequest.java
@@ -8,5 +8,5 @@ import lombok.Data;
 public class SkillUpgradeRequest {
     @NotNull(message = "技能ID不能为空")
     @Min(value = 1, message = "技能ID必须大于0")
-    private Long playerSkillId;
+    private Integer playerSkillId;
 }
